Improved theme support (see desc.)
Themes previously consisted of one css file that could be switched in
the settings, now themes consist of a complete folder. The weather
forecast svg is now also theme-dependant and uses a css in the theme
subfolder. Still not ideal, changes now have to be made in two files at
the same time.

This might be a good idea though, as it will allow to create entirely
different forecast graphics based on themes