Fix rendering error by splitting SVG and Website CSS in separate files
Rendering error occurs on Chrome, when the same css is referenced from
both SVG and webpage because of external font definitions.
This causes the fonts to be loaded twice and sometimes the font will not
render at all in Chrome (reload frantically to reproduce).

By splitting in two files the fonts are only loaded once and the bug disappears.
Edge and Opera seemingly had no problems..